requirements.txt
torch>=2.0.0
transformers>=4.30.0
txtai>=6.0.0
python-docx>=0.8.11
requests>=2.31.0
numpy>=1.24.0
scikit-learn>=1.3.0
pydantic>=2.0.0
networkx>=3.0
pyyaml>=6.0
gliner>=0.0.2